feat: make permission config usable in client, move server utils to server dir
