+++ apps/website/src/server/utils/auth.ts
@@ -1,6 +1,8 @@
 import type { Session } from "next-auth";
 
-import { env } from "../../env/server.mjs";
+import { env } from "@/env/server.mjs";
+import type { PermissionConfig } from "@/config/permissions";
+import { getRolesForUser } from "@/server/db/users";
 
 export function getSuperUserIds() {
   return env.SUPER_USER_IDS.split(",").map((id) => id.trim());
@@ -17,4 +19,26 @@ export function checkIsSuperUserId(id?: string) {
 
 export function checkIsSuperUserSession(session: Session | null) {
   return checkIsSuperUserId(session?.user?.id);
+}
+
+export async function checkPermissions(
+  permissionConfig: PermissionConfig,
+  userId?: string
+) {
+  if (!userId) {
+    return false;
+  }
+
+  const isSuperUser = checkIsSuperUserId(userId);
+  if (isSuperUser) {
+    return true;
+  }
+
+  if (!permissionConfig.requiresSuperUser && permissionConfig.requiredRole) {
+    return (await getRolesForUser(userId)).includes(
+      permissionConfig.requiredRole
+    );
+  }
+
+  return false;
 }
